requirements.txt
requests
anthropic
openai
python-dotenv
